Allow font size to a be double (1 digit) instead of an integer.
A previous commit make the font size an integer.  Per Christoph Feck
there are valid reasons to allow x.y font sizes.  This patch restores
that ability.

An update/fix to ebc594cdf643eae09c89afe9674ed20ea39d91d7

CCMAIL: christoph@maxiom.de

+++ src/EditProfileDialog.h
@@ -118,7 +118,7 @@ private slots:
     void silenceSecondsChanged(int);
 
     // appearance page
-    void setFontSize(int pointSize);
+    void setFontSize(double pointSize);
     void setFontInputValue(const QFont&);
     void setAntialiasText(bool enable);
     void setBoldIntense(bool enable);
